extract the theme
